requirements.txt
psycopg2-binary==2.9.9
django-environ==0.11.2
django==4.2.13 #5.0.6
djoser==2.2.2
djangorestframework==3.15.1
django-debug-toolbar==4.4.2
psycopg2-binary==2.9.9
django-environ==0.11.2
python-dateutil==2.9.0
gunicorn==22.0.0
boto3==1.34.145
django-storages==1.14.2
fcm-django==2.2.1
python-dotenv==1.0.1